refactor(deps): remove composer/composer package dependency
The composer/composer package dependency was only used to read the
contents of the extra section of a project's composer.json file.
This commit replaces this logic by standard PHP functions.

As a result, the composer/composer dependency has been removed
which results in potentially 21 less dependencies when installing
this package as a dependency.

Closes: #103